price_update_tests/requirements.txt
flask
requests
pandas
openpyxl
google-auth
google-auth-oauthlib
google-api-python-client
google-cloud-storage

